Extracted some modules into the components
- Extracted crc, msp_ltm_serial and tcp_server into separate components
- Updated the CMakeLists.txt file
- Moved crc32 related functions to the crc component
- Deleted unused component.mk file
- Removed licencse header from some files
