+++ components/crc/CMakeLists.txt
@@ -1,0 +1,2 @@
+idf_component_register(SRCS "crc.c"
+                       INCLUDE_DIRS ".")
