+++ components/crc/crc.c
@@ -1,23 +1,4 @@
-/*
- *   This file is part of DroneBridge: https://github.com/DroneBridge/ESP32
- *
- *   Copyright 2017 Wolfgang Christl
- *
- *   Licensed under the Apache License, Version 2.0 (the "License");
- *   you may not use this file except in compliance with the License.
- *   You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- *   Unless required by applicable law or agreed to in writing, software
- *   distributed under the License is distributed on an "AS IS" BASIS,
- *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *   See the License for the specific language governing permissions and
- *   limitations under the License.
- *
- */
-
-#include "db_crc.h"
+#include "crc.h"
 
 /**
  * This function is part of Cleanflight & iNAV.
@@ -57,4 +38,45 @@ uint8_t crc8_dvb_s2(uint8_t crc, unsigned char a)
 uint8_t crc8_dvb_s2_table(uint8_t crc, unsigned char a)
 {
     return (uint8_t) (crc_dvb_s2_table[(crc ^ a)] & 0xff);
+}
+
+/**
+ * @brief Calculate CRC32 of a string
+ * 
+ * @param crc Initial CRC value (0 if not known)
+ * @param buf Buffer containing string
+ * @param len Length of the string
+ * @return uint32_t CRC32 value
+ */
+uint32_t calc_crc32(uint32_t crc, unsigned char *buf, size_t len) {
+    int k;
+    crc = ~crc;
+    while (len--) {
+        crc ^= *buf++;
+        for (k = 0; k < 8; k++)
+            crc = crc & 1 ? (crc >> 1) ^ 0xedb88320 : crc >> 1;
+    }
+    return ~crc;
+}
+
+/**
+ * @brief Check CRC32 of a string
+ * 
+ * @param buf Buffer containing string
+ * @param msg_length Length of the string
+ * @return 1 if CRC is good else 0
+ */
+int crc_ok(uint8_t *buf, int msg_length) {
+    uint32_t c_crc = calc_crc32((uint32_t) 0, buf, (size_t) (msg_length - 4));
+    uint8_t crc_bytes[4];
+    crc_bytes[0] = c_crc;
+    crc_bytes[1] = c_crc >> 8;
+    crc_bytes[2] = c_crc >> 16;
+    crc_bytes[3] = c_crc >> 24;
+    if ((crc_bytes[0] == buf[msg_length - 4]) && (crc_bytes[1] == buf[msg_length - 3]) &&
+        (crc_bytes[2] == buf[msg_length - 2]) && (crc_bytes[3] == buf[msg_length - 1])) {
+        return 1;
+    } else {
+        return 0;
+    }
 }
