+++ components/msp_ltm_serial/CMakeLists.txt
@@ -1,0 +1,3 @@
+idf_component_register(SRCS "msp_ltm_serial.c"
+                       INCLUDE_DIRS "."
+                       PRIV_REQUIRES crc)
