+++ components/tcp_server/CMakeLists.txt
@@ -1,0 +1,3 @@
+idf_component_register(SRCS "tcp_server.c"
+                       INCLUDE_DIRS "."
+                       PRIV_REQUIRES esp_event log lwip)
